[Graphics][URP][Bugfix]Fixed reflection probes being in reverse order for forward plus
Error introduced in this PR: https://github.cds.internal.unity3d.com/unity/unity/pull/39057
Reflection probes are currently in the reverse order we need it for forward+ which this PR undo.
This PR also adds a test case for reflection probes and lights using the mobile light and reflection probe limit. To catch this error in the future as well as other issues.