Graphics/SRP/RPF - [UUM 60100] Add missing CreateSkyboxRendererList in RG API
As noted by a Unity forum user, Render Graph API does not provide a utility function to create a skyboxRendererList while the other specific native rendering (ui overlay, wireframe, gizmos...) are covered by the API. 

This PR adds the missing API to let users create skyboxRendererList through Render Graph API.

URP DrawSkyboxPass has been updated to use the new API with the Render Graph path. It had to be refactored slightly as RendererList cannot be provided to user before the Render Graph execution time, only the RendererListHandle.

RenderGraphTests has been updated to allow testing with a ScriptableRenderContext through a custom Render Pipeline. This is necessary to test this part of the Render Graph API and it will be useful in a near future to test Render Graph workflow in its Execute step or over multiple frames. Note that I don't think we should need a ScriptableRenderContext ready at this level of the API and I would prefer LegacyRendererLists to follow the classic RendererList workflow but this is outside the scope of this bug fix.