Enable ErrorMaterial test for BatchRenderingGroup in URP
The test that we run in this PR verifies that the magenta material is applied when using BRG on URP. It is a graphic test using image comparison. 

This PR is similar to this one[PR](https://github.cds.internal.unity3d.com/unity/unity/pull/44096), except that it targets URP and also contains a fix to the test scene. 

Just like for the HDRP version, it contains changes that prevent the ErrorMaterial test to fail because of shader compilation errors in the logs (which are expected). 

More specifically:

1 - Ignore failing messages using UTF LogAssert API (see GraphicsTests.cs)

2 - Disable scene in build profile when running the build job in Yamato (see urp-brg.metafile)

UTR does not allow to ignore failing messages (only compilation errors), this is why I opted to disable the scene in the BuildSettings.

Additionally, the ErrorMaterial scene was not correctly set up to show  the error material when DOTS Multiinstancing is not enabled in the shader. Specifically, the changes made to the scene are:

1 - Bound the RenderBRG game object to the RenderBRG.cs script
2 - Made the cubes game object children of the RenderBRG game object