[JIRA # UUM-65155] Fixing stencil compiler bug for subsurface scattering on scarlett
Fix coarse stencil bug computation race condition caused by a compiler bug in DXC.
The issue manifests itself as flickering tiles on sub surface scatter pixels. This issue is caused because the DXC compiler does not make the branches using a groupThreadId vector comparison a scalar. Changing this to use the group index instead (preflattened) triggers the compiler to recognize the branch as a scalar instead.