[HDRP] Fix issues with RTGI and RTR discrepency when using high contrasted IBL
This PR is fixing discrepency observe when using an IBL with strong local lighting and using SSGI/SSR versus RTGI/RTR in quality, performance and mixed mode (see the JIRA for more details).

The PR is removing the clamped value for the "clamp" parameter of RTR and RTGI allowing to put any value and is changing the default from 1.0 to 100.0.
Users observing noise with raytracing will be require to adjust those value to improve the quality, pottentially causing a decrease in intensity, but at least by default users will not complain that the lighting is dull and not understanding why.

The PR also update the fallback of the RTGI and SSGI effect to fallback on ambient probe rather than Sky cubemap texture for less noisy result.