[VFX] Fix editor crash when deleting textures that are used for rendering in VFX
In some scenarios, it was possible to crash the editor if textures, models or graphics buffers were deleted when an active VFX was using them.

Not all the ways to destroy these objects will cause a crash, it depends on the path taken. For instance, deleting texture using "Del" key would work fine, but deleting it using right click + "Delete" would crash.

This was due to a cached pointer, used for optimization, that was not invalidated. In VFX there are 3 different command lists, 1 running during update (safe, because it is when the pointers are cached). And 2 running during the rendering. If some object is deleted between VFX manager update and rendering, it would cause the crash.

The solution was to keep track of which effects need to reload the objects cache and do that during the first rendering command list (PerCamera).

On top of that, we removed some unused values from output update that made the crash more likely to happen with 6 way lighting.