Fixed GRD invalid state when changing rendererMode
If data changes that could affect the result of IsGPUResidentDrawerSupportedBySRP(), we must reinitialize the GPU Resident Drawer. For URP, we are checking if the rendererMode is Forward+ for all renderers, else GPU Resident Drawer is not used. 
Previously when rendererMode was changed, we did nothing, which could result in GPU Resident Drawer being enabled/disabled when it shouldn't be.