Fix Default Resource Initialization for APV
APV was using the TextureXR::GetBlackTexture3D to fallback to a default resource if the main runtime APV resources were not yet initialized.
However in URP, TextureXR::Initialize is never invoked so these default resources never exist, resulting in some console warning spam.

Instead of initializing the TextureXR resources, I switched the usage to CoreUtils::blackVolumeTexture. I examined the usage of these APV resources on GPU side and confirmed that there is no actual need for an XR-friendly format for sampling the probe data.