Remove Apply and Reset mip bias in HDRP
Replace the costly multiple mip bias CBuffer updates dring the rendering by a hardcoded define `SUPPORT_GLOBAL_MIP_BIAS` in shaders.

The idea is that we push the correct global mip bias once in the shader variables CBuffer and then the shader passes can decide whether or not to use it by enabling the define `SUPPORT_GLOBAL_MIP_BIAS` before including `ShaderVariables.hlsl`.
All passes that can sample user textures should have this option enabled.

Additionally, after post processes passes are forced with a mip bias of 0. This matches the previous behaviour but it's still wrong as it ignores the material bias in the camera data.

Additionally removes 12 uploads of the global ShaderVariables CBuffer during the frame.