DOCG-3358,4907,5119,4897 Add render graph system documentation
Add introduction and task pages for RenderGraph.

**This PR is dependent on https://github.cds.internal.unity3d.com/unity/unity/pull/41475**

This PR focuses on an initial set of user concepts and tasks we agreed with Felipe, @oleksandr-kokoshyn and @sophia-lloyd. The following is out of scope for this PR:

- Any new documentation about Compatibility Mode.
- OnBeginRenderGraphFrame, OnEndRenderGraphFrame
- ConfigureInput (as it's not specifically a Render Graph API). It is mentioned and included briefly in code samples though.)
- AllowGlobalStateModification
- Write a combined RG and non-RG pass.
- RenderTextureDescriptor (as it's not specifically a Render Graph API) - it is mentioned and included in code samples though.)
- SetGlobalTexture - as SetGlobalTextureAfterPass is preferred
- Compute shaders
- Cross-pipeline

Please note too that I've followed our [guidance on code samples](https://docs-style-guide.unity.com/content-types/code-examples/create-code-examples) to prioritize readability over performance or completeness. So code samples might not be 'best practice' so it's easier for users to understand the core concept.