[APV] Add missing API to overwrite result of APV baking
Allows users to implement their own lightmapper (see Bakery on the asset store)
All actual public API newly exposed are in [this commit](fa1ea59e47cfe481a0154b68eabe8aff6c661436), so you can review them quickly

The idea is that we expose an API to override each baking backend, namely
- virtual offset computation
- sky occlusion & shading direction
- probe sh & validity

If any of those is not overriden, the default implementation will be used.
The baking backend for SH supports being run in a thread. This may not be used by users but it's supported.

The benefit of this method is that everything works with our existing UI and workflows, so it's supposed to integrate nicely in a project if an asset store overrides baking.

To allow bypassing the lightmapper and still use APV, some APIs have also been added to bake probe volume only