SpeedTree 9 - Workflow improvements
Improve the `SpeedTree 9` importer workflow in multiple cases:
- When the `render pipeline` is changed (e.g from HDRP to built-in and/or opposite). It's now cleaner.
- When the `pipeline asset` is changed.
- When the `default shader` is changed.

It now retriggers a reimport of the st9 assets for all these cases. The last 2 were not supported correctly.

It also removes the dependency on the direct path of the assets, so there's less chance to break anything (e.g the asset is being moved, we don't have to change the path manually in the importer anymore).