Enable Foveated Rendering during Render Pass Initialization
This PR addresses the issue raised here: https://jira.unity3d.com/browse/XR-2400. 

Previously, `SetRenderPassAttachments` was called before enabling foveation, resulting in edge cases where render passes that expect to have foveation data may be initialized without it. 

This PR makes it so foveation is properly set before the `SetRenderPassAttachments` call is made. This aligns the behavior of the ScriptableRenderer with what is on trunk (Unity 2023). 

https://github.cds.internal.unity3d.com/unity/unity/blob/trunk/Packages/com.unity.render-pipelines.universal/Runtime/ScriptableRenderer.cs#L1667-L1669

There's no specific PR to backport as the state of trunk was a result of several unrelated changes.

This change is needed because Unity 2022 is the current stable developer ecosystem for VisionOS