Fix core samples dependencies errors with Input system package
Thanks to @will feedbacks, we discovered there's some errors when importing HDRP samples into a brand new project. 
The probleme is due to the project player settings input set to New Input System but not having the input system package installed on the project. (IMHO, thoses two things should be linked, but that's another discussion). 
This fix is in two parts: 
- The first part has already been done, and was just about including input system package by default in the hdrp blank template manifest. More info here. More info on @kristin-stock [PR](https://github.cds.internal.unity3d.com/unity/rm-package-promotion/pull/1348#event-2079656)
- The second part is to prevent errors to happen even if users uninstalls input package. One solution is adding an assembly reference asset that defines a keyword if the input package is installed. Then, on each script using input system, we test both of these keywork (if the player settings input is set to new input system and if the input system package is installed)