[VFX] Split rendering batches for exposed meshes
When we allowed using exposed objects with instancing, we missed one case, using a exposed mesh with mesh particles.

The fix contains 2 parts:
- Identifying which CPU expressions can split instancing batches. Currently just mesh and submeshMask, but easy to add more
- Supporting indirect draw rendering with different meshes

To test, just expose a mesh and plug it directly into the mesh slot of a particle mesh output. Then, create many instances and use different meshes for some instances.