[VFX] UI Optimization to improve performance
[Slack channel](https://unity.enterprise.slack.com/archives/C06D7PTSJ9H)
We recently noticed that switching tab to a big VFX Graph asset could take several seconds.
After investigation, we were told that's because the whole content is being recreated from scratch.
In the scope of Unity6 nothing is planned in UIToolkit, nor in GraphView to improve this behavior.
So the idea here is to reduce at the minimum the number of UI elements used to build our Graph.

In the process of this UI optimization, there I found several small design issues or even bugs that I could easily fix in the process of optimization (a large part of the work was to tweak uss anyway).

So far, the performance improvements measure give the following numbers:
**~ 30% less UI elements**
**~ 40% faster when switching tab**    

See captures below (beta 8)
![ui-profiling-comparison-annotated](https://media.github.cds.internal.unity3d.com/user/4003/files/85c2a95f-bce6-4cc8-82b4-a0ceabe1e80b)    
    
For the record, here are some small UI improvements
![ui-comparison](https://media.github.cds.internal.unity3d.com/user/4003/files/5d8a2b75-9bf1-47f2-a1c3-4c3a8fa47389)


Some other improvements:
- Node's expand collapse button is nicer and has hover visual feedback
- Fixed issue when dragging a link from an "untyped" port (Add, Multiply ...) and using the ALT modifier key