Fix Render Graph Viewer warnings on HDRP due to duplicate pass names
Fix https://jira.unity3d.com/browse/UUM-65820

When capturing data for Render Graph Viewer, we use the RG pass name as key in the dictionary that stores the script file name and line information required for "jump to pass source in IDE" feature. If duplicate pass names are found, a warning is printed to let user know that this mapping will be ambiguous.

In the case of this bug, the helper functions `BeginProfilingSampler` and `EndProfilingSampler` enqueue render passes with a fixed name, leading to multiple duplicate entries and therefore warnings. However, as these passes also call `builder.GenerateDebugData(false)`, they are not visible in RG Viewer in the first place and we can safely ignore them.

Note that the new code added is only ever active during the frame where RG Viewer is opened or the refresh button is clicked, so there is no performance impact.