Add missing LightCookies keyword to screenspace decal projector pass
Fixes https://jira.unity3d.com/browse/UUM-54404, where screen space decals weren't taking into account light cookies.

The code to sample cookies is already present in the relevant shader (a shader made with shadergraph decal template), but the keyword for light cookies, `_LIGHT_COOKIES` is not declared. The fix is simply to add the keyword to the relevant pass descriptor.

The keyword will automatically be set earlier in the render loop, in `LightCookiesManager.Setup`.