[RPW] Switch to default and current render pipeline
We complete switching to the newest API in GraphicsSettings.

GraphicsSettings.renderPipelineAsset API will be deprecated and it's replaced by defaultRenderPipeline. 
We should use GraphicsSettings.defaultRenderPipeline or GraphicsSettings.currentRenderPipeline.

Incorrect API was used in some places. It supposed to rely on the current pipeline instead of default as we may not have a default but we could still have RP in the Quality Settings.

1. Scene Render mode window could have incorrect height
2. Avatar previews could be handled incorrectly
3. Light component have an issue in a multiple RP projects
4. ScreenSpaceAmbientOcclusion may not work when URP is deferred in a multiple RP projects
5. Enlighten Manager may not work if you don't set RP asset in the GraphicsSettings


https://jira.unity3d.com/browse/UUM-66743