[Yamato] Fix & re-enable unstable IsAssetPreviewCorrect test
Fix unstable test: https://jira.unity3d.com/browse/UUM-59403 which was disabled earlier due to instability.
 
The test was unstable because it sometimes took too long and timed out. The instability was actually introduced by this PR https://github.cds.internal.unity3d.com/unity/unity/pull/41268 that tracks timeouts correctly, meaning that the highly varying runtimes of this test (usually 100-300sec but up to 700 sec sometimes) were not being enforced before. After the UTF 1.4.2 release the timeouts started correctly failing the test, and the test became unstable as a consequence.

This PR fixes the issue by refactoring it: it used to be a edit mode test that manually enters playmode (triggering domain reload) before running the test logic. It was originally done this way to more closely follow the repro steps where asset preview bugs would happen. But given the long runtime, it makes more sense to just implement it as a playmode test instead, so that we don't need to enter playmode separately which is time-consuming. As a result, the test is now passing reliably and only taking ~5sec.

cc @alex-vazquez - There was a separate issue with URP_Foundation project where the playmode test assembly was being excluded from editor, meaning that we had tests that were not being run when they should have been. These tests have now been fixed and enabled too.