UUM-65454 MaxSubpass discrepancy with RenderPassSetup and NativePassCompiler
JIRA: https://jira.unity3d.com/browse/UUM-65454

When merging subpasses together in a NativePass, we don't check if there will be more than 8 native render subpasses. This PR adds a check for this when verifying if a certain graph pass can be merged into a Native Pass, as well as a test. If merging the graph pass would result in too many native subpasses, we simply create a new Native Pass