[RPW] Multiple fixes for the different Editors
https://jira.unity3d.com/browse/UUM-62739
https://jira.unity3d.com/browse/UUM-62751
https://jira.unity3d.com/browse/UUM-62981

UUM-62739 Fixes for the Volume Component. ⚠️ Only part 1 done. 2 and 3 should be ignored for the Testing
![image](https://media.github.cds.internal.unity3d.com/user/5932/files/4cb6d1e1-eb8b-4a50-8f7d-4e0579e8d049)

UUM-62751 Fixed a tooltip for the Material UI.
![image](https://media.github.cds.internal.unity3d.com/user/5932/files/95cff8b2-2d21-4f0f-bbbb-c5d30dd4183e)

UUM-62981
Removed not only 3 dots menu but also warning and error buttons which appear when there's an empty object field. Text box is still there and below the list.