[HDRP] Fix clouds not affected by intensity multiplier
Currently the intensity multiplier doesn't affect clouds, so that means, when trying to swap shadow maps on Sun/Moon directionnal light, even if both celestials bodies have an intensity multiplier at 0, we still see a pop ([video](https://gyazo.com/7bea0971a9e9e8a3e237c4a8bd3f1c2d)). 

This PR fixes it for both clouds and volumetric clouds. 

In the case we, one day, support multiple directionnal light with clouds (it has a cost), we'll need to remove this so that it works as the PBR sky (which supports multiple lights)