Prefab preview thumbnails are not being rendered when a Project uses HDRP
In 2023.2, assets preview thumbnails are not rendered properly sometimes, they are gray. It's not the case in `2023.3.0a7`, it works perfectly fine.

It works because of the introduced changes in this PR: https://github.cds.internal.unity3d.com/unity/unity/pull/34816
Specifically the changes for the `HDRenderPipeline.cs`. Removing the globalSettings null check in 2023.2 fixes the issue. 

I don't know this code area, this is definitely more a question for @alex-vazquez 
I have to jump in vacation this week, I apologize for that, but wanted to create this PR so I can help the next person investigating this issue. I don't mind resuming my work on it if we wanna wait a week!