[Port] [2023.2] Fix for the preview rendering issue (no default sky ambient)
In the initial rendering of the preview in the inspector window, the ambient color of the default sky may not be ready (black ambient color), due to the lag of GPU readback of the SH coefficients. The ambient color is not updated until the preview is updated by some triggers like hovering the cursor over the inspector.
This PR fixes the issue by using the hard-coded value of the default sky's SH.