fix UUM-68261 URP ignores resourceData.cameraColor when set in a user pass AfterRenderPostProcessing
fix [UUM-68261](https://jira.unity3d.com/browse/UUM-68261) URP ignores resourceData.cameraColor when set in a user pass AfterRenderPostProcessing

Update the RenderGraph blit with material sample to cover this scenario, and demonstrate how to use the ResourceData to avoid a copy blit back to the resource. This update also fixes bugs in the sample.