Rendering Debugger - GPU Resident Drawer - Adding feedback on the debug panel why some settings are not applied.
Fix this issue:

https://jira.unity3d.com/browse/UUM-67130

Extracted the information why the GPU Resident drawer is not supported and added that feedback into the Rendering Debugger.

Also, now the panel shows the warning and hides modifiying the GPU Resident Drawer if it is not enabled.

The `RecreateIfNeeded`, is need to be handled per Render, as ther IsEnabled depends on settings that come directly from GraphicsSettings, and not from Render Pipeline Graphics settings, Therfore, when you changed something the GPU Resident drawer was not recreated.

Added capabilities to the debug message box, to display new message.