Fix preview camera drawing grid on top of meshes with URP RG & depth priming
Fix https://jira.unity3d.com/browse/UUM-55171

This PR fixes issue where when Render Graph is active, animation preview window draws the grid on top of the preview object. This is because the grid is rendered in BiRP after rendering the preview object, and the depth buffer is not passed correctly when using URP RG. This only happens with Forward & depth priming active.

The fix has two parts:
1) UniversalRendererRenderGraph.cs: Added missing check to enable FinalCopyDepthPass. This is the same bugfix as was done earlier to fix grid depth composition: https://github.cds.internal.unity3d.com/unity/unity/pull/16980 - however the original fix was only made on non-RG codepath, so we need the same fix on the RG codepath.

2) CopyDepthPass.cs: Doing the fix above, we start to get this warning from C++: `Attempting to render to a depth only surface with no dummy color attachment` (see [GfxDevice.cpp](https://github.cds.internal.unity3d.com/unity/unity/blob/trunk/Runtime/GfxDevice/GfxDevice.cpp#L1545)). To fix this, the same fix is applied to preview camera as was done for scene view camera here: https://github.cds.internal.unity3d.com/unity/unity/pull/36665

After these two fixes, the BiRP code that renders the grid gets the URP-rendered depth buffer, and the grid is rendered correctly.

Before:
![image](https://media.github.cds.internal.unity3d.com/user/3380/files/b122b988-1f4e-4f1b-9f26-24593e9f7b62)

After:
![image](https://media.github.cds.internal.unity3d.com/user/3380/files/1e0c018b-64d2-4ad2-85d2-45453774e5b1)