[Graphics]Adding mip level and slices to color and depth attachment for RG
This PR is related to the [URP-2105](https://jira.unity3d.com/browse/URP-2105) jira epic.
The purpose of this PR is to add missing functionality by adding mip level and slices to RG's  SetInputAttachment and SetRenderAttachment functions.

This PR also introduces helper functions for Copy and Blit as well as a test to ensure the newly added functionality and the helper functions work as expected.

There is also a fix for invalid texture description for imported textures