Fix Profilers.EndSample error with custom render passes calling Submit()
Fix https://jira.unity3d.com/browse/UUM-69066

This bug happens when an user calls `Submit()` inside `ScriptableRenderPass.Execute()`. This is causing the profiler marker order to be ill-formed.

Here is more detailled explanation of what happens. URP `ScriptableRenderer.Execute` function looks like this:

```
void ScriptableRendererExecute()
{
    using (new ProfilingScope(cmd, profilingExecute))
    {
        /* ... */

        foreach (ScriptableRenderPass renderPass in renderPassBlock)
        {
            renderPass.Execute();
        }

        /* ... */
    }
}
```
And `ProfilingScope` ctor/dtor look like this:

```
ProfilingScope(CommandBuffer cmd, ProfilingSampler sampler)
{
    cmd.BeginSample(sampler); 
    sampler.inlineSampler.Begin();
}

void Dispose()
{
    cmd.EndSample(sampler);
    sampler.inlineSampler.End();
}
```
So here is what happens with this setup when Submit() is called inside ScriptableRenderPass.Execute() :
```
void ScriptableRendererExecute()
{
    cmd.BeginSample(sampler); // Deferred profiler_begin() which will be called in Submit()
    sampler.inlineSampler.Begin(); // Calls profiler_begin() immediately

    // Inside renderPass.Execute()
    context.Submit(); // Calls the deferred profiler_begin()

    cmd.EndSample(sampler); // Deferred profiler_end() which will be called in Submit()
    sampler.inlineSampler.End(); / Calls profiler_end() immediately -> Error. The previous profiler_begin() call has no corresponding profiler_end()
}
```
Since we have no way to know in advance whether a render pass will call `Submit()` or not, we can't know if it is fine to use command buffer based profiling scope without causing errors. So the fix proposed here is to simply not use that when there is a risk that Submit() is called inside of the profiling scope. We can simply stick to normal immediate profiler markers instead.