Fix viewport flickering when multi-frame rendering API is used with certain parameters
Our user was getting flickering when using accumulation motion blur with certain parameters:
https://jira.unity3d.com/browse/UUM-59115

The root cause of this was that we were not properly handling the case that the first sample in the Monte Carlo accumulation has zero weight. I've re-arranged the formula to output zero in this case. 

Path tracing without motion blur always uses a weight of one for all frames, so it should not be affected. @pieterjan-bartels if you think there is an issue let me know.
 
I've also fixed an error in the example script in the documentation: it was capturing a screenshot in two places, only one is needed.