[HDRP] Fix Transmission not being affected by shadow from Volumetric Clouds.
While developping arctic demo, we noticed that the transmission in our icy moutains was still glowing eventhough they were in the volumetric clouds shadows, this was a huge discrepancy in lighting which is now eaisly fixed by this PR thanks to @adrien-tocqueville snippet.

This wasn't the case for cloud layer because it's using the cookie of the directionnal light. 


![c5784cf2073607debfbe678b852b052e](https://media.github.cds.internal.unity3d.com/user/1764/files/7819d363-b832-4a2c-856b-e9d0d4599b8e)

Fix hard earth shadow
![image](https://media.github.cds.internal.unity3d.com/user/2154/files/5adb6416-6dff-4ad4-9afc-bd154048475d)
![image](https://media.github.cds.internal.unity3d.com/user/2154/files/64573645-fe94-48e1-bfa2-de8aa97af159)