fixing URP UniversalResourceData resource setters that should not be public
Fixing URP frame resource setters that should not be public. This avoids many kinds of user errors and bugs due to unexpected behavior. It also minimizes the issue of frame resource handles and global texture reference becoming out-of-sync. This is a partial fix to https://jira.unity3d.com/browse/UUM-69948

This is new API that landed in 23.3 but that accidently made setters public that should not be, since it's incorrect user behavior to set these. This PR adds guardrails to guide users better. Since this is new API, and a very advanced use case, this should only impact a handful of users at most.